+++ R/ggscatterstats.R
@@ -14,6 +14,7 @@
 #' @param xlab Label for `x` axis variable.
 #' @param ylab Label for `y` axis variable.
 #' @param line.colour Colour for the regression line.
+#' @param line.size Size for the regression line.
 #' @param marginal Decides whether `ggExtra::ggMarginal()` plots will be
 #'   displayed; the default is `TRUE`.
 #' @param marginal.type Type of marginal distribution to be plotted on the axes
@@ -85,25 +86,26 @@
 # defining the function
 ggscatterstats <-
   function(data = NULL,
-             x,
-             y,
-             xlab = NULL,
-             ylab = NULL,
-             line.colour = "blue",
-             marginal = TRUE,
-             marginal.type = "histogram",
-             width.jitter = NULL,
-             height.jitter = NULL,
-             xfill = "orange",
-             yfill = "green",
-             centrality.para = NULL,
-             type = "pearson",
-             results.subtitle = NULL,
-             title = NULL,
-             caption = NULL,
-             maxit = 500,
-             k = 3,
-             messages = TRUE) {
+           x,
+           y,
+           xlab = NULL,
+           ylab = NULL,
+           line.size = 1.5,
+           line.colour = "blue",
+           marginal = TRUE,
+           marginal.type = "histogram",
+           width.jitter = NULL,
+           height.jitter = NULL,
+           xfill = "#E69F00",
+           yfill = "#D55E00",
+           centrality.para = NULL,
+           type = "pearson",
+           results.subtitle = NULL,
+           title = NULL,
+           caption = NULL,
+           maxit = 500,
+           k = 3,
+           messages = TRUE) {
     # if data is not available then don't display any messages
     if (is.null(data)) {
       messages <- FALSE
@@ -112,11 +114,9 @@ ggscatterstats <-
     # preparing a dataframe out of provided inputs
     if (!is.null(data)) {
       # preparing labels from given dataframe
-      lab.df <- colnames(dplyr::select(
-        .data = data,
-        !!rlang::enquo(x),
-        !!rlang::enquo(y)
-      ))
+      lab.df <- colnames(dplyr::select(.data = data,
+                                       !!rlang::enquo(x),
+                                       !!rlang::enquo(y)))
       # if xlab is not provided, use the variable x name
       if (is.null(xlab)) {
         xlab <- lab.df[1]
@@ -135,10 +135,8 @@ ggscatterstats <-
     } else {
       # if vectors are provided
       data <-
-        base::cbind.data.frame(
-          x = x,
-          y = y
-        )
+        base::cbind.data.frame(x = x,
+                               y = y)
     }
 
     ######################################## statistical labels ######################################################
@@ -159,7 +157,7 @@ ggscatterstats <-
 
         c <-
           stats::cor.test(
-            formula = ~x + y,
+            formula = ~ x + y,
             data = data,
             method = "pearson",
             alternative = "two.sided",
@@ -208,7 +206,7 @@ ggscatterstats <-
         # note that stats::cor.test doesn't give degress of freedom; it's calculated as df = (no. of pairs - 2)
         c <-
           stats::cor.test(
-            formula = ~x + y,
+            formula = ~ x + y,
             data = data,
             method = "spearman",
             alternative = "two.sided",
@@ -230,8 +228,11 @@ ggscatterstats <-
           )) %>%
           tibble::as_data_frame(x = .) %>%
           dplyr::select(.data = ., estimate) %>%
-          dplyr::summarize(.data = ., low = quantile(estimate, 0.05 / 2),
-                           high = quantile(estimate, 1 - 0.05 / 2))
+          dplyr::summarize(
+            .data = .,
+            low = quantile(estimate, 0.05 / 2),
+            high = quantile(estimate, 1 - 0.05 / 2)
+          )
 
         # preparing the label
         stats_subtitle <-
@@ -279,11 +280,9 @@ ggscatterstats <-
 
         # getting confidence interval for rho
         c_ci <-
-          stats::confint.default(
-            object = MASS_res,
-            parm = "scale(x)",
-            level = 0.95
-          )
+          stats::confint.default(object = MASS_res,
+                                 parm = "scale(x)",
+                                 level = 0.95)
 
         # preparing the label
         stats_subtitle <-
@@ -318,9 +317,8 @@ ggscatterstats <-
               df = summary(MASS_res)$df[2],
               # degrees of freedom are always integer
               pvalue = ggstatsplot::specify_decimal_p(sfsmisc::f.robftest(MASS_res)$p.value[[1]],
-                k,
-                p.value = TRUE
-              )
+                                                      k,
+                                                      p.value = TRUE)
             )
           )
         # displaying the details of the test that was run
@@ -340,26 +338,20 @@ ggscatterstats <-
 
     # preparing the scatterplotplot
     plot <-
-      ggplot2::ggplot(
-        data = data,
-        mapping = ggplot2::aes(
-          x = x,
-          y = y
-        )
-      ) +
+      ggplot2::ggplot(data = data,
+                      mapping = ggplot2::aes(x = x,
+                                             y = y)) +
       ggplot2::geom_point(
         size = 3,
         alpha = 0.5,
-        position = position_jitter(
-          width = width.jitter,
-          height = height.jitter
-        ),
+        position = position_jitter(width = width.jitter,
+                                   height = height.jitter),
         na.rm = TRUE
       ) +
       ggplot2::geom_smooth(
         method = "lm",
         se = TRUE,
-        size = 1.5,
+        size = line.size,
         colour = line.colour,
         na.rm = TRUE
       ) +
@@ -380,7 +372,8 @@ ggscatterstats <-
 
     if (is.null(centrality.para)) {
       plot <- plot
-    } else if (isTRUE(centrality.para) || centrality.para == "mean") {
+    } else if (isTRUE(centrality.para) ||
+               centrality.para == "mean") {
       plot <- plot +
         ggplot2::geom_vline(
           xintercept = mean(data$x),
@@ -423,14 +416,10 @@ ggscatterstats <-
           p = plot,
           type = marginal.type,
           size = 5,
-          xparams = base::list(
-            fill = xfill,
-            col = "black"
-          ),
-          yparams = base::list(
-            fill = yfill,
-            col = "black"
-          )
+          xparams = base::list(fill = xfill,
+                               col = "black"),
+          yparams = base::list(fill = yfill,
+                               col = "black")
         )
 
       ################################################### messages ##########################################################